tests/wpt/css-tests/css-flexbox-1_dev/html/flexbox_computedstyle_justify-content-flex-end.htm

<!DOCTYPE html>
<html><head><title>flexbox | computed style | justify-content: flex-end</title>
<link href="http://opera.com" rel="author" title="Opera Software">
<link href="http://www.w3.org/TR/css-flexbox-1/#propdef-justify-content" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags">
<style>
body {
	color: red;
	justify-content: flex-end;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("justify-content"),
	          "flex-end");
});
</script>
</body></html>
